Move compute_vmap to core/_utils
