+++ dynamiqs/_utils.py
@@ -2,8 +2,6 @@ from __future__ import annotations
 
 from typing import Any
 
-import jax
-import numpy as np
 from jax import numpy as jnp
 from jaxtyping import Array
 
@@ -43,27 +41,3 @@ def bexpect(O: Array, x: Array) -> Array:
     if isket(x):
         return jnp.einsum('ij,bjk,kl->b', dag(x), O, x)  # <x|O|x>
     return jnp.einsum('bij,ji->b', O, x)  # tr(Ox)
-
-
-def compute_vmap(
-    f: callable,
-    cartesian_batching: bool,
-    is_batched: list[bool],
-    out_axes: list[int | None],
-) -> callable:
-    if any(is_batched):
-        if cartesian_batching:
-            # iteratively map over the first axis of each batched argument
-            idx_batched = np.where(is_batched)[0]
-            # we apply the successive vmaps in reverse order, so that the output
-            # batched dimensions are in the correct order
-            for i in reversed(idx_batched):
-                in_axes = [None] * len(is_batched)
-                in_axes[i] = 0
-                f = jax.vmap(f, in_axes=in_axes, out_axes=out_axes)
-        else:
-            # map over the first axis of all batched arguments
-            in_axes = list(np.where(is_batched, 0, None))
-            f = jax.vmap(f, in_axes=in_axes, out_axes=out_axes)
-
-    return f
