+++ dynamiqs/core/_utils.py
@@ -2,6 +2,8 @@ from __future__ import annotations
 
 from typing import get_args
 
+import jax
+import numpy as np
 from jax import numpy as jnp
 from jaxtyping import ArrayLike
 
@@ -31,3 +33,27 @@ def get_solver_class(
             f' solver types: {supported_str}).'
         )
     return solvers[type(solver)]
+
+
+def compute_vmap(
+    f: callable,
+    cartesian_batching: bool,
+    is_batched: list[bool],
+    out_axes: list[int | None],
+) -> callable:
+    if any(is_batched):
+        if cartesian_batching:
+            # iteratively map over the first axis of each batched argument
+            idx_batched = np.where(is_batched)[0]
+            # we apply the successive vmaps in reverse order, so that the output
+            # batched dimensions are in the correct order
+            for i in reversed(idx_batched):
+                in_axes = [None] * len(is_batched)
+                in_axes[i] = 0
+                f = jax.vmap(f, in_axes=in_axes, out_axes=out_axes)
+        else:
+            # map over the first axis of all batched arguments
+            in_axes = list(np.where(is_batched, 0, None))
+            f = jax.vmap(f, in_axes=in_axes, out_axes=out_axes)
+
+    return f
